# backend/openapi.yaml
openapi: 3.0.3
info:
  title: AgentRQ API
  version: 1.0.0
  description: |
    OpenAPI specification for the AgentRQ backend HTTP API.

    ## Authentication

    Every route except those under `/auth` requires the `at` session cookie,
    set by one of the login flows. There is no bearer-token path: the frontend
    and the desktop app both address the API with same-origin relative URLs and
    let the cookie ride along.

    ## Conventions

    - All request and response JSON uses `camelCase`.
    - IDs are base62-encoded monoflakes rendered as strings, never numbers.
    - Most payloads are wrapped in a single named key — `{"task": {...}}`,
      `{"workspaces": [...]}` — rather than returned bare.

    ## Errors

    Two shapes are in use. Handlers that map a domain error return the
    structured envelope (`ErrorResponse`); several that validate inline return
    a flat `{"error": "message"}` (`SimpleError`). Each operation below names
    the one it returns.
servers:
  - url: /api/v1
...
paths:
...
  /workspaces/{id}/skills:
    parameters:
      - $ref: '#/components/parameters/WorkspaceId'
    get:
      tags: [Workspaces]
      summary: The account's skills, as this workspace sees them
      description: >
        Every skill of the account, by name, each with `workspaceEnabled`:
        whether it is on in this workspace (its parent, for a fork). The
        workspace's agent sees a skill only when it is on here and `enabled`
        for the account. With `q`, only skills whose name or description
        contains it, ignoring case. `files` is omitted.
      parameters:
        - name: q
          in: query
          required: false
          schema: { type: string, minLength: 3, maxLength: 256 }
          description: Text to find in a skill's name or description. Leave it out to list every skill.
        - name: limit
          in: query
          required: false
          schema: { type: integer, minimum: 0, maximum: 100 }
          description: How many skills to return; larger values are capped at 100. Leave it out, or 0, to return every match.
        - name: offset
          in: query
          required: false
          schema: { type: integer, minimum: 0, default: 0 }
          description: How many matches to skip.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  skills:
                    type: array
                    items: { $ref: '#/components/schemas/Skill' }
                  total:
                    type: integer
                    description: How many skills match in all, across every page.
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/InvalidPayload' }
  /workspaces/{id}/skills/{name}:
    parameters:
      - $ref: '#/components/parameters/WorkspaceId'
      - $ref: '#/components/parameters/SkillName'
    get:
      tags: [Workspaces]
      summary: One skill as this workspace sees it, with its list of files
      description: The files are listed by path and size, without content.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  skill: { $ref: '#/components/schemas/Skill' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/InvalidPayload' }
    patch:
      tags: [Workspaces]
      summary: Turn a skill on or off in this workspace
      description: >
        On, this workspace's agent sees the skill, as long as it is on for the
        account too. Off, the skill stays in the account and in every other
        workspace it is on in. A fork's switch is its parent's.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [enabled]
              properties:
                enabled: { type: boolean }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  skill: { $ref: '#/components/schemas/Skill' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/InvalidPayload' }
...
  /skills:
    get:
      tags: [Skills]
      summary: Search the account's skills
      description: >
        Every skill of the account, by name, each with the workspaces it is on
        in. With `q`, only skills whose name or description contains it,
        ignoring case. `files` is omitted.
      parameters:
        - name: q
          in: query
          required: false
          schema: { type: string, minLength: 3, maxLength: 256 }
          description: Text to find in a skill's name or description. Leave it out to list every skill.
        - name: limit
          in: query
          required: false
          schema: { type: integer, minimum: 0, maximum: 100 }
          description: How many skills to return; larger values are capped at 100. Leave it out, or 0, to return every match.
        - name: offset
          in: query
          required: false
          schema: { type: integer, minimum: 0, default: 0 }
          description: How many matches to skip.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  skills:
                    type: array
                    items: { $ref: '#/components/schemas/Skill' }
                  total:
                    type: integer
                    description: How many skills match in all, across every page.
        '422': { $ref: '#/components/responses/InvalidPayload' }
  /skills/import:
    post:
      tags: [Skills]
      summary: Import skills from a public GitHub repository
      description: >
        Downloads the repository once and imports every skill in it. A plugin
        manifest, `.agentrq/plugin.json` or else `.muse-plugin/plugin.json`,
        says which directories are skills; without one, every directory with a
        `SKILL.md` is. A skill keeps its Markdown files, except a repository's
        own README, CLAUDE, AGENTS and similar meta files, and any other file
        its kept files reference, directly or in turn. Anything left out is
        listed in `skipped` with the reason, rather than failing the import.
        A repository too large to download whole imports nothing: its skills
        come back in `candidates`, and the import is asked again with the
        wanted ones in `skills`, which are then read file by file. Every
        skill imported belongs to the account, and is turned on in each
        workspace of `workspaceIds`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [url]
              properties:
                url:
                  type: string
                  example: https://github.com/obra/superpowers
                  description: >
                    `https://github.com/<owner>/<repo>`, optionally ending in
                    `.git`, `/tree/<ref>` or `/tree/<ref>/<path>`.
                overwrite:
                  type: boolean
                  default: false
                  description: >
                    Replace the account's skill that has the same name. It keeps
                    the workspaces it is on in, and whether it is on at all.
                skills:
                  type: array
                  maxItems: 256
                  items: { type: string }
                  example: [ship, careful]
                  description: >
                    Import only these skills, by their directory in the
                    repository as `candidates` gives it.
                workspaceIds:
                  type: array
                  items: { type: string }
                  description: >
                    Workspaces (base62) to turn each imported skill on in. A
                    fork stands for its parent.
      responses:
        '200':
          description: What was imported, and what was left out
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SkillImportReport' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422':
          description: Not a GitHub repository link, or one GitHub does not serve publicly
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '502':
          description: GitHub could not be reached, or would not list a repository too large to import whole
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /skills/{name}:
    parameters:
      - $ref: '#/components/parameters/SkillName'
    get:
      tags: [Skills]
      summary: One skill, with its list of files
      description: The files are listed by path and size, without content.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  skill: { $ref: '#/components/schemas/Skill' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/InvalidPayload' }
    patch:
      tags: [Skills]
      summary: Turn a skill on or off for every workspace
      description: >
        A skill that is off stays in the account, and on in its workspaces,
        but no agent sees it: `searchSkills` leaves it out and `loadSkill`
        does not find it. Skills are on when created or imported, and a
        re-import keeps one off.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [enabled]
              properties:
                enabled: { type: boolean }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  skill: { $ref: '#/components/schemas/Skill' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/InvalidPayload' }
    delete:
      tags: [Skills]
      summary: Delete a skill and its files from the account
      description: Every workspace it was on in loses it.
      responses:
        '204': { description: Deleted }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/InvalidPayload' }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: at
      description: Access token cookie, set by any of the `/auth` login flows.
  parameters:
    WorkspaceId:
      name: id
      in: path
      required: true
      schema: { type: string }
      description: Workspace ID (base62).
...
    SkillName:
      name: name
      in: path
      required: true
      schema: { type: string, maxLength: 64, pattern: '^[a-z0-9]+(-[a-z0-9]+)*$' }
      description: The skill's name, unique among the skills a workspace can use.
  responses:
    InvalidPayload:
      description: Invalid request payload
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    NotFound:
      description: Not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  schemas:
    ErrorResponse:
      type: object
      description: The structured envelope returned by handlers that map a domain error.
      properties:
        error:
          type: object
          properties:
            message: { type: string }
            code: { type: integer }
...
    Skill:
      type: object
      description: >
        A playbook agents load for a kind of task: a `SKILL.md` and the files it
        references. It belongs to the account, its name unique there, and is
        turned on in workspaces one by one. Agents read it over MCP as
        `skill://<name>/<path>`.
      properties:
        id: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        name: { type: string, maxLength: 64 }
        description:
          type: string
          maxLength: 1024
          description: From the SKILL.md frontmatter; what an agent matches on to load it.
        sourceType: { type: string, enum: [manual, github] }
        sourceRepo: { type: string, description: '`owner/repo`, for an imported skill.' }
        sourceRef: { type: string }
        sourceCommit: { type: string, description: The commit imported; empty when GitHub did not say. }
        sourcePath: { type: string, description: The skill's directory in the repository. }
        locallyModified:
          type: boolean
          description: An imported skill that has been changed since.
        fileCount: { type: integer, description: Files including SKILL.md. At most 256. }
        totalBytes: { type: integer }
        enabled:
          type: boolean
          description: The account-wide switch. Off, no agent sees the skill, whichever workspaces it is on in.
        workspaceIds:
          type: array
          items: { type: string }
          description: The workspaces the skill is on in, whose agents see it while `enabled` is true.
        workspaceEnabled:
          type: boolean
          description: Given only when the skill is read through a workspace, and then whether it is on there.
        files:
          type: array
          description: Present only when one skill was requested by name.
          items: { $ref: '#/components/schemas/SkillFile' }
    SkillFile:
      type: object
      properties:
        path: { type: string, maxLength: 255 }
        sizeBytes:
          type: integer
          description: At most 98304 for SKILL.md and 65536 for any other file.
        updatedAt: { type: string, format: date-time }
        content:
          type: string
          description: Present only when the file itself was requested.
        url:
          type: string
          format: uri
          readOnly: true
          description: |
            Public link to the file, readable with no session: the bucket when
            skills are kept in S3, `<base URL>/storage/skills/...` otherwise.
    SkillImportReport:
      type: object
      properties:
        imported:
          type: array
          items:
            type: object
            properties:
              name: { type: string }
              fileCount: { type: integer }
              totalBytes: { type: integer }
        skipped:
          type: array
          description: Skills and single files left out, each with the reason.
          items:
            type: object
            properties:
              name: { type: string }
              path: { type: string }
              reason: { type: string }
        candidates:
          type: array
          description: >
            Present only when the repository is too large to import whole and
            no skills were chosen: the skills to choose from, with the reason
            one cannot be imported when that is already known.
          items:
            type: object
            properties:
              name: { type: string }
              path: { type: string }
              sizeBytes: { type: integer, description: The size of its SKILL.md }
              reason: { type: string }
        sourceRepo: { type: string }
        sourceRef: { type: string }
        sourceCommit: { type: string }
...
security:
  - sessionCookie: []